Switch root constants into root CBV (camera constants, shadow constants), separate gViewProj

+++ frame_debug.cpp
@@ -77,9 +77,9 @@ namespace udsdx
 		SetTextColor(m_hMemDC, RGB(255, 255, 255));
 
 		std::wstring text = std::format(L"{:4} fps", static_cast<int>(1.0f / time.deltaTime));
-		TextOutW(m_hMemDC, 0, 0, text.c_str(), text.size());
+		TextOutW(m_hMemDC, 0, 0, text.c_str(), (int)text.size());
 		text = std::format(L"{:6.1} ms", time.deltaTime * 1000);
-		TextOutW(m_hMemDC, 0, 16, text.c_str(), text.size());
+		TextOutW(m_hMemDC, 0, 16, text.c_str(), (int)text.size());
 
 		SelectBitmap(m_hDC, m_memBitmap);
 		BitBlt(m_hDC, 0, 0, WindowWidth, WindowHeight, m_hMemDC, 0, 0, SRCCOPY);
